+++ src/pages/OnBoarding/APPLEAuthRedirectPage.tsx
@@ -3,17 +3,18 @@ import { useAppDispatch } from '@store/app/hooks';
 import { useNavigate } from 'react-router-dom';
 
 function APPLEAuthRedirectPage() {
+  console.log('apple callback 진입');
   const dispatch = useAppDispatch();
   const navigate = useNavigate();
 
   useEffect(() => {
     document.addEventListener('AppleIDSignInOnSuccess', (data) => {
-      console.log('handle successful response');
+      console.log('handle success');
       console.log('data:', data);
     });
     document.addEventListener('AppleIDSignInOnFailure', (error) => {
-      console.log('handle error');
-      console.log('error: ', error);
+      console.error('handle error');
+      console.error('error: ', error);
     });
   }, []);
